+++ cojopy/__init__.py
@@ -8,7 +8,7 @@ from .cojopy import COJO
 
 __author__ = """Jianhua Wang"""
 __email__ = "jianhua.mert@gmail.com"
-__version__ = '0.0.2'
+__version__ = "0.0.2"
 
 
 logging.basicConfig(