+++ cojopy/cli.py
@@ -42,8 +42,9 @@ def main(
 
 @app.command(name="slct", help="Conditional selection of SNPs.")
 def slct(
-    sumstats: str = typer.Argument(..., help="Path to the summary statistics file."),
+    sumstats: str = typer.Argument(..., help="Path to the summary statistics file. The file should have the same columns as COJO input: SNP, A1, A2, b, se, p, freq, N"),
     ld_matrix: str = typer.Argument(..., help="Path to the LD matrix file."),
+    ld_freq: str = typer.Option(None, "--ld-freq", "-f", help="Path to the LD frequency file. The file should have the following columns: SNP, freq. Use freq in sumstats if ld_freq is not provided."),
     p_cutoff: float = typer.Option(5e-8, "--p-cutoff", "-p", help="P-value cutoff."),
     collinear_cutoff: float = typer.Option(
         0.9, "--collinear-cutoff", "-c", help="Collinearity cutoff."
@@ -51,17 +52,55 @@ def slct(
     window_size: int = typer.Option(
         10000000, "--window-size", "-w", help="Window size."
     ),
+    maf_cutoff: float = typer.Option(0.01, "--maf-cutoff", "-m", help="Minor allele frequency cutoff."),
+    diff_freq_cutoff: float = typer.Option(
+        0.05, "--diff-freq-cutoff", "-d", help="Difference in minor allele frequency cutoff."
+    ),
     output: str = typer.Option(..., "--output", "-o", help="Path to the output file."),
 ):
     """Perform conditional selection of SNPs using COJO algorithm."""
     c = COJO(
-        p_cutoff=p_cutoff, collinear_cutoff=collinear_cutoff, window_size=window_size
+        p_cutoff=p_cutoff,
+        collinear_cutoff=collinear_cutoff,
+        window_size=window_size,
+        maf_cutoff=maf_cutoff,
+        diff_freq_cutoff=diff_freq_cutoff,
     )
-    sumstats_df = pd.read_csv(sumstats, sep="\t")
-    ld_matrix_array = np.loadtxt(ld_matrix)
-    cojo_result = c.conditional_selection(sumstats_df, ld_matrix_array)
-    cojo_result.to_csv(output, sep="\t", index=False)
+    cojo_result = c.conditional_selection(sumstats, ld_matrix, ld_freq)
+    cojo_result.to_csv(output, sep="\t", index=False, float_format="%.6g")
 
+
+@app.command(name="joint", help="Joint analysis of SNPs.")
+def joint(
+    sumstats: str = typer.Argument(..., help="Path to the summary statistics file. The file should have the same columns as COJO input: SNP, A1, A2, b, se, p, freq, N"),
+    ld_matrix: str = typer.Argument(..., help="Path to the LD matrix file."),
+    ld_freq: str = typer.Option(None, "--ld-freq", "-f", help="Path to the LD frequency file. The file should have the following columns: SNP, freq. Use freq in sumstats if ld_freq is not provided."),
+    extract_snps: str = typer.Option(None, "--extract-snps", "-e", help="Path to the file containing the SNPs to extract. Each SNP should be in a new line."),
+    p_cutoff: float = typer.Option(5e-8, "--p-cutoff", "-p", help="P-value cutoff."),
+    collinear_cutoff: float = typer.Option(
+        0.9, "--collinear-cutoff", "-c", help="Collinearity cutoff."
+    ),
+    window_size: int = typer.Option(
+        10000000, "--window-size", "-w", help="Window size."
+    ),
+    maf_cutoff: float = typer.Option(0.01, "--maf-cutoff", "-m", help="Minor allele frequency cutoff."),
+    diff_freq_cutoff: float = typer.Option(
+        0.05, "--diff-freq-cutoff", "-d", help="Difference in minor allele frequency cutoff."
+    ),
+    output: str = typer.Option(..., "--output", "-o", help="Path to the output file."),
+):
+    """Perform joint analysis of SNPs using COJO algorithm."""
+    logger = logging.getLogger("COJO")
+    c = COJO(
+        p_cutoff=p_cutoff,
+        collinear_cutoff=collinear_cutoff,
+        window_size=window_size,
+        maf_cutoff=maf_cutoff,
+        diff_freq_cutoff=diff_freq_cutoff,
+    )
+    joint_result = c.run_joint_analysis(sumstats, ld_matrix, extract_snps, ld_freq)
+    logger.info("Joint analysis complete. Writing results to %s", output)
+    joint_result.to_csv(output, sep="\t", index=False, float_format="%.6g")
 
 if __name__ == "__main__":
     app(main)