+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.3] (2025-03-20)
+
+* Fixed bug in joint statistics calculation.
+* Added tests for conditional selection and joint analysis.
+
 ## [0.0.2] (2025-03-09)
 
 * Conditional selection with LD matrix.